infra: Add test for rhel vm with d1 instance type
