+++ tests/infrastructure/instance_types/conftest.py
@@ -21,7 +21,7 @@ from utilities.artifactory import (
 )
 from utilities.constants import (
     CONTAINER_DISK_IMAGE_PATH_STR,
-    DATA_SOURCE_STR,
+    OS_FLAVOR_RHEL,
     OS_FLAVOR_WIN_CONTAINER_DISK,
     TIMEOUT_20MIN,
     Images,
@@ -162,15 +162,29 @@ def windows_vm_for_dedicated_cpu(request, unprivileged_client, namespace, latest
         vm_instance_type=VirtualMachineClusterInstancetype(
             client=unprivileged_client, name=request.param["instance_type_name"]
         ),
-        vm_preference=VirtualMachineClusterPreference(
-            client=unprivileged_client,
-            name=py_config["latest_windows_os_dict"][DATA_SOURCE_STR].replace("win", "windows."),
-        ),
+        vm_preference_infer=True,
         data_volume_template=data_volume_template_with_source_ref_dict(
             data_source=latest_windows_data_source,
         ),
         os_flavor=OS_FLAVOR_WIN_CONTAINER_DISK,
         disk_type=None,
+    ) as vm:
+        vm.start()
+        yield vm
+
+
+@pytest.fixture()
+def rhel_vm_for_dedicated_cpu(unprivileged_client, namespace, latest_rhel_data_source):
+    with VirtualMachineForTests(
+        client=unprivileged_client,
+        name="rhel-d1-vm",
+        namespace=namespace.name,
+        vm_instance_type=VirtualMachineClusterInstancetype(client=unprivileged_client, name="d1.large"),
+        vm_preference_infer=True,
+        data_volume_template=data_volume_template_with_source_ref_dict(
+            data_source=latest_rhel_data_source,
+        ),
+        os_flavor=OS_FLAVOR_RHEL,
     ) as vm:
         vm.start()
         yield vm
